% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erickmath.R
\name{erickmath}
\alias{erickmath}
\title{erickmath}
\usage{
erickmath(number)
}
\arguments{
\item{number}{This is the number that the function takes as an
argument}
}
\value{
This function returns the number 0 no matter what
}
\description{
This is a function that will return 0 for anything you
give it
}
